Добавил генерацию dto

+++ src/main/java/ru/spgorshkov/user/dao/UserRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.spgorshkov.user.entity.UserEntity;
 
 @Repository
-public interface UserRepository extends PagingAndSortingRepository<UserEntity, Long> {
+public interface UserRepository extends PagingAndSortingRepository<UserEntity, Integer> {
 }
